Crate zwreec

Twee-to-Zcode compile library.

Zwreec is a compiler for interactive fiction in the Twee format (created by the Twine software) to Z-Machine instructions (Zcode) which can be run with interpreters like frotz.

This library was developed as part of the course "Softwareprojekt Übersetzerbau" in 2015.

Requirements and Usage

To build Zwreec from Source, you will need to have both Rust 1.1.0 and Cargo installed on your system. You can download the Rust binarys on their website, by using your system's package manager or by running this in your shell:

$ curl -sSf https://static.rust-lang.org/rustup.sh | sh

Cargo should be installed alongside current Rust binaries.

The library uses a fork of rustlex to do lexical analysis.

To use zwreec in your project you can add it as a dependency to your Cargo.toml.

[dependencies.zwreec]
git = "https:://github.com/Drakulix/zwreec"

Then you can use it in your crate root.

extern crate zwreec;

Example

The following example is itself a full working command line compiler. It is a simpler version of the Reference Binary Implementation.

extern crate zwreec;

use std::env;
use std::error::Error;
use std::fs::File;
use std::path::Path;

fn main() {
    let mut args: Vec<String> = env::args().collect();

    if args.len() != 2 { panic!("Need exactly one input file!"); }

    let cfg = zwreec::config::Config::default_config();
    let mut input = match File::open(Path::new(&args[1])) {
        Ok(file) => file,
        Err(why) => { panic!("Couldn't open input: {}", Error::description(&why)); }
    };
    let mut output = match File::create(Path::new("a.z8")) {
        Ok(file) => file,
        Err(why) => { panic!("Couldn't open output: {}", Error::description(&why)); }
    };

    zwreec::compile(cfg, &mut input, &mut output);
}

Reference Binary Implementation

Zwreecs Github-Repository contains a reference binary implementation that uses this library and provides a simple command line interface to compile Twee files.

To build the binary, you need the Rust version as outlined in Requirements and Usage. Then you can build the binary using

$ cargo build

The resulting binary can be found at target/debug/zwreec.

Logging

Zwreec uses the logging facade provided by log. The reference binary implementation of Zwreec also includes an implementation of the Log trait.

Modules

backend
The backend module is responsible for processing the abstract syntax tree (AST) generated by the frontend and generating an output file that can be run with a Z-Machine interpreter of your choosing.

config
Configuration for the zwreec compiler.

frontend
Processes Twee files and builds an abstract syntax tree.

utils
Utility code for more ituitive and shorter code.

Macros

error_force_panic!
Report an error and panic in any case

error_panic!
Report an error

Functions

compile
Compiles a Twee Input to Zcode

test_library
Run internal library tests.