Module zwreec::backend

The backend module is responsible for processing the abstract syntax tree (AST) generated by the frontend and generating an output file that can be run with a Z-Machine interpreter of your choosing.

It works by using the codegen module to generate the code.

Example

This example shows how you can use the backend to generate Z-Code. The "ast" variable already contains an iterator over zwreec::frontend::ast::ASTNodes generated by the frontend

// Create a configuration
let cfg = zwreec::config::Config::default_config();

// Output vector
let vec: Vec<u8> = vec![];
let mut output = Cursor::new(vec);

// generate code
zwreec::backend::codegen::generate_zcode(&cfg, ast, &mut output);

Modules

codegen
The codegen module is for the creation of Z-Code from an AST. The generate_zcode function traverses the AST and emits Z-Code along the way by using the functions provided by the Z-Code backend module. The Z-Code data is then written to the output. See the module level documentation for more information on how to use the backend.

zcode
The zcode module contains all functionality to deal with the Z-Code related stuff.