Module zwreec::backend::codegen

The codegen module is for the creation of Z-Code from an AST. The generate_zcode function traverses the AST and emits Z-Code along the way by using the functions provided by the Z-Code backend module. The Z-Code data is then written to the output. See the module level documentation for more information on how to use the backend.

Structs

CodeGenManager
The manager that contains a lot of state for the code generation.

IdentifierProvider
A generator for unique IDs.

SymbolTable
The symbol table.

Enums

CodeGenError
All the errors that can occur during code generation.

Functions

function_random
This generates code for the function random(from, to) -> zcode op_random(0, range).

gen_zcode
Generate Z-Code based on the ASTNode and its children.

generate_zcode
Create Codegen state and generate Z-Code from the specified AST passage iterator.