Function zwreec::backend::codegen::generate_zcode

pub fn generate_zcode<W: Write, I: Iterator<Item=ASTNode>>(cfg: &Config, ast: I, output: &mut W)

Create Codegen state and generate Z-Code from the specified AST passage iterator.

Panics

This panics when an error occurs and the force option in the config is not set. It also panics even when that option is set for invalid expressions, missing passages and other unrecoverable errors.