Struct zwreec::backend::codegen::CodeGenManager

pub struct CodeGenManager<'a> {
    pub cfg: &'a Config,
    pub ids_if: IdentifierProvider,
    pub ids_expr: IdentifierProvider,
    pub ids_link_var_set: IdentifierProvider,
    pub visited_passages: HashSet<String>,
    pub required_passages: Vec<String>,
    pub symbol_table: SymbolTable,
    pub format_state: FormattingState,
    pub is_silent: bool,
    pub is_nobr: bool,
}

The manager that contains a lot of state for the code generation.

Fields

cfg

The zwreec config

ids_if

The ID provider for if labels

ids_expr

The ID provider for expressions

ids_link_var_set

The ID provider for set variable operations

visited_passages

The passages already processed by Codegen

required_passages

All passages that are linked to (including Start)

symbol_table

The symbol table

format_state

The current formatting options

is_silent

Is this inside a silent tag? (no output)

is_nobr

Is this inside a nobr tag? (no line breaks)

Methods

impl<'a> CodeGenManager<'a>

fn new(cfg: &'a Config) -> CodeGenManager<'a>

Creates a new CodeGenManager.

fn new_temp_var_vec() -> Vec<u8>

Creates a vector with 13 temporary IDs (15 is the count of allowed max local variables in Z-Code and two variables are reserved for other purposes).

fn is_temp_var(var: &Variable) -> bool

Tells whether a variable is a temporary (true) or global variable (false).

fn validate_passages(&self)

Checks for Twee invariants (Start passage must exist, all linked passages must exist).