Struct zwreec::backend::codegen::SymbolTable

pub struct SymbolTable {
    // some fields omitted
}

The symbol table.

Methods

impl SymbolTable

fn new() -> SymbolTable

Creates a new symbol table.

fn insert_new_symbol(&mut self, symbol: String, t: Type)

Inserts a symbol into the table, assigning a new id.

fn is_known_symbol(&self, symbol: &String) -> bool

Checks if the symbol is already existent in the table.

fn get_symbol_id(&self, symbol: &String) -> Variable

Returns the id for a given symbol.

Panics

This panics when the symbol is unkown.

fn get_and_add_symbol_id(&mut self, symbol: String) -> Variable

Returns the id for a given symbol.

Checks is the symbol is known, otherwise adds it silently as type None.

fn get_symbol_type(&self, symbol: &String) -> Type

Returns the Type of the specified symbol.

Panics

This panics when the symbol map does not contain the symbol.

fn has_var_id(&self, id: u8) -> bool

Checks if the symbol table contains a variable with the specified id.

fn get_symbol_type_by_id(&self, id: u8) -> Type

Returns the Type of the symbol specified by the id.

Panics

This panics if the symbol table does not contain the symbol id.