Struct zwreec::backend::zcode::zbytes::Bytes

pub struct Bytes {
    pub bytes: Vec<u8>,
}

A struct that holds an array of bytes and provides some convenience functions.

Fields

bytes

The underlying data

Methods

impl Bytes

fn len(&self) -> usize

Returns the length of the byte array.

fn write_byte(&mut self, byte: u8, index: usize)

Writes the byte (u8) to the index specified.

If the vector isn't large enough it fills everything up to the index with zeros.

fn append_byte(&mut self, byte: u8)

Appends a byte to the end of the data.

fn write_u16(&mut self, value: u16, index: usize)

Writes a u16 in two bytes with the correct byte-order for the Z-Machine at the specified index.

fn append_u16(&mut self, value: u16)

Appends a u16 to the end of the data.

fn write_bytes(&mut self, bytes: &[u8], to_index: usize)

Writes multiple bytes at the specified index.

fn append_bytes(&mut self, bytes: &[u8])

Appends an array of bytes at the end of the data.

fn write_zero_until(&mut self, index: usize)

Fills everything with zeros until but not including the index.

=> [index-1] == 0; [index] == nil;

fn print(&self)

Prints the underlying byte array