Module zwreec::backend::zcode::zfile

The zfile module contains functionality to create a Z-Code file.

This contains all code related to operand handling, Z-Code instruction generation (except for most op-codes) and all of shared structs and code used by codegen and other modules.

Reexports

pub use super::zbytes::Bytes;

pub use super::ztext;

pub use super::ee::routine_easteregg;

pub use super::op;

Structs

Constant
An integer constant.

FormattingState
A formatting type.

LargeConstant
A signed 16-bit integer constant.

Variable
A variable.

Zfile
The definition of a Z-Code file.

Zjump
A jump.

Zlabel
A label.

Zstring
A string.

Enums

ArgType
Types of possible arguments.

JumpType
Zcode has the jump-types:

Operand
There are three Operands in Z-Code: Variables, SmallConsts and LargeConsts.

Type
A variable type.

ZOP
ZOP: Z-Code op code and pseudo op-code representation.