Enum zwreec::backend::zcode::zfile::ZOP

pub enum ZOP {
    PrintUnicode {
        c: u16,
    },
    PrintUnicodeVar {
        var: Variable,
    },
    PrintUnicodeStr {
        address: Operand,
    },
    PrintChar {
        var: Variable,
    },
    Print {
        text: String,
    },
    PrintNumVar {
        variable: Variable,
    },
    PrintVar {
        variable: Variable,
    },
    PrintPaddr {
        address: Operand,
    },
    PrintAddr {
        address: Operand,
    },
    PrintOps {
        text: String,
    },
    Call1N {
        jump_to_label: String,
    },
    Call2NWithAddress {
        jump_to_label: String,
        address: String,
    },
    Call2NWithArg {
        jump_to_label: String,
        arg: Operand,
    },
    Call1NVar {
        variable: u8,
    },
    Call2S {
        jump_to_label: String,
        arg: Operand,
        result: Variable,
    },
    CallVNA2 {
        jump_to_label: String,
        arg1: Operand,
        arg2: Operand,
    },
    CallVNA3 {
        jump_to_label: String,
        arg1: Operand,
        arg2: Operand,
        arg3: Operand,
    },
    CallVSA2 {
        jump_to_label: String,
        arg1: Operand,
        arg2: Operand,
        result: Variable,
    },
    CallVSA3 {
        jump_to_label: String,
        arg1: Operand,
        arg2: Operand,
        arg3: Operand,
        result: Variable,
    },
    CallVS2A5 {
        jump_to_label: String,
        arg1: Operand,
        arg2: Operand,
        arg3: Operand,
        arg4: Operand,
        arg5: Operand,
        result: Variable,
    },
    Routine {
        name: String,
        count_variables: u8,
    },
    Label {
        name: String,
    },
    Newline,
    SetColor {
        foreground: u8,
        background: u8,
    },
    SetColorVar {
        foreground: u8,
        background: u8,
    },
    SetTextStyle {
        bold: bool,
        reverse: bool,
        monospace: bool,
        italic: bool,
    },
    StoreVariable {
        variable: Variable,
        value: Operand,
    },
    StoreW {
        array_address: Operand,
        index: Variable,
        variable: Variable,
    },
    StoreB {
        array_address: Operand,
        index: Variable,
        variable: Variable,
    },
    StoreBOperand {
        array_address: Operand,
        index: Operand,
        operand: Operand,
    },
    LoadBOperand {
        array_address: Operand,
        index: Operand,
        variable: Variable,
    },
    PushVar {
        variable: Variable,
    },
    PullVar {
        variable: Variable,
    },
    Inc {
        variable: u8,
    },
    Dec {
        variable: u8,
    },
    Ret {
        value: Operand,
    },
    JE {
        operand1: Operand,
        operand2: Operand,
        jump_to_label: String,
    },
    JNE {
        operand1: Operand,
        operand2: Operand,
        jump_to_label: String,
    },
    JL {
        operand1: Operand,
        operand2: Operand,
        jump_to_label: String,
    },
    JLE {
        operand1: Operand,
        operand2: Operand,
        jump_to_label: String,
    },
    JG {
        operand1: Operand,
        operand2: Operand,
        jump_to_label: String,
    },
    JGE {
        operand1: Operand,
        operand2: Operand,
        jump_to_label: String,
    },
    Random {
        range: Operand,
        variable: Variable,
    },
    ReadChar {
        local_var_id: u8,
    },
    ReadCharTimer {
        local_var_id: u8,
        timer: u8,
        routine: String,
    },
    AddTypes {
        operand1: Operand,
        operand2: Operand,
        tmp1: Variable,
        tmp2: Variable,
        save_variable: Variable,
    },
    Add {
        operand1: Operand,
        operand2: Operand,
        save_variable: Variable,
    },
    Sub {
        operand1: Operand,
        operand2: Operand,
        save_variable: Variable,
    },
    Mul {
        operand1: Operand,
        operand2: Operand,
        save_variable: Variable,
    },
    Div {
        operand1: Operand,
        operand2: Operand,
        save_variable: Variable,
    },
    Mod {
        operand1: Operand,
        operand2: Operand,
        save_variable: Variable,
    },
    Or {
        operand1: Operand,
        operand2: Operand,
        save_variable: Variable,
    },
    And {
        operand1: Operand,
        operand2: Operand,
        save_variable: Variable,
    },
    Not {
        operand: Operand,
        result: Variable,
    },
    Jump {
        jump_to_label: String,
    },
    LoadW {
        array_address: Operand,
        index: Variable,
        variable: Variable,
    },
    SetCursor {
        line: u8,
        col: u8,
    },
    SetCursorOperand {
        row: Operand,
        col: Operand,
    },
    UpdateCursorPos,
    GetCursor {
        store_addr: Operand,
    },
    EraseWindow {
        value: i8,
    },
    EraseLine,
    SetVarType {
        variable: Variable,
        vartype: Type,
    },
    CopyVarType {
        variable: Variable,
        from: Operand,
    },
    GetVarType {
        variable: Variable,
        result: Variable,
    },
    Quit,
}

ZOP: Z-Code op code and pseudo op-code representation.

These structs contain instructions to generate Z-Machine op-codes. To generate the op-codes see backend::zcode::zfile::Zfile::emit.

Variants

PrintUnicode

Prints a unicode character.

Fields

c

PrintUnicodeVar

Prints the unicode character in the specified variable. var contains a 16 bit variable that contains the character.

Fields

var

PrintUnicodeStr

Prints the unicode string at the address specified.

Fields

address

PrintChar

Print a ZSCII character.

Fields

var

Print

Print a ZSCII string.

Fields

text

PrintNumVar

Print a numeric variable.

Fields

variable

PrintVar

Print a variable.

Fields

variable

PrintPaddr

Print the ZSCII string at the packed address address.

Fields

address

PrintAddr

Print the ZSCII string at the large address specified.

Fields

address

PrintOps

Generate print op-codes for the specified string.

This generates ZSCII op-codes for the supported characters and prints unicode characters separately.

Fields

text

Call1N

1OP: Call a routine with one argument.

Fields

jump_to_label

Call2NWithAddress

2OP: Call a routine with one argument (at the specified label) and throws the result away.

Fields

jump_to_label

address

Call2NWithArg

2OP: Call a routine with one argument and store the result in arg.

Fields

jump_to_label

arg

Call1NVar

1OP: Call a routine with one variable argument.

Fields

variable

Call2S

2OP: Call a routine with one argument and store the return value in result.

Fields

jump_to_label

arg

result

CallVNA2

VAROP: Call a routine with two arguments and throw result away.

Fields

jump_to_label

arg1

arg2

CallVNA3

VAROP: Call a routine with three arguments and throw result away.

Fields

jump_to_label

arg1

arg2

arg3

CallVSA2

VAROP: Call a routine with two arguments and store result in result.

Fields

jump_to_label

arg1

arg2

result

CallVSA3

VAROP: Call a routine with three arguments and store result in result.

Fields

jump_to_label

arg1

arg2

arg3

result

CallVS2A5

VAROP with types-byte: Call a routine with five arguments and store the return value in result.

Fields

jump_to_label

arg1

arg2

arg3

arg4

arg5

result

Routine

Declares a Z-Routine

Fields

name

count_variables

Label

Declares a label.

This is only used internally to reference different parts of the generated file.

Fields

name

Newline

Prints a newline.

SetColor

Sets the foreground and background color to constants specified.

Fields

foreground

background

SetColorVar

Sets the foreground and background color to the variables with the IDs specified.

Fields

foreground

background

SetTextStyle

Set text style to bold, reverse (inverse colors), monospace and italic.

Fields

bold

reverse

monospace

italic

StoreVariable

Store the value in value to the variable.

Fields

variable

value

StoreW

Store the word in variable at array_address + index.

Fields

array_address

index

variable

StoreB

Store the byte in variable at array_address + index.

Fields

array_address

index

variable

StoreBOperand

Store the byte in operand at array_address + index.

Fields

array_address

index

operand

LoadBOperand

Load the byte at array_address + index into variable.

Fields

array_address

index

variable

PushVar

Push a variable on the stack.

Fields

variable

PullVar

Pull a variable from the stack.

Fields

variable

Inc

Increment a variable by 1: variable += 1.

Fields

variable

Dec

Decrement a variable by 1: variable -= 1.

Fields

variable

Ret

Return from the Z-Routine with the value in the specified Operand.

Fields

value

JE

Jump if operand1 == operand2.

Fields

operand1

operand2

jump_to_label

JNE

Jump if operand1 != operand2.

Fields

operand1

operand2

jump_to_label

JL

Jump if operand1 < operand2.

Fields

operand1

operand2

jump_to_label

JLE

Jump if operand1 <= operand2.

Fields

operand1

operand2

jump_to_label

JG

Jump if operand1 > operand2.

Fields

operand1

operand2

jump_to_label

JGE

Jump if operand1 >= operand2.

Fields

operand1

operand2

jump_to_label

Random

Store a random number between 1 and range in variable.

Fields

range

variable

ReadChar

Read a character from standard input in the variable.

Fields

local_var_id

ReadCharTimer

Read a character from standard input in the variable or time out after timer / 10 seconds elapsed.

Fields

local_var_id

timer

routine

AddTypes

Helper function to add two values according to their types.

Fields

operand1

operand2

tmp1

tmp2

save_variable

Add

Add two values: save_variable = operand1 + operand2.

Fields

operand1

operand2

save_variable

Sub

Subtract two values: save_variable = operand1 - operand2.

Fields

operand1

operand2

save_variable

Mul

Multiply two values: save_variable = operand1 * operand2.

Fields

operand1

operand2

save_variable

Div

Divide two values: save_variable = operand1 / operand2.

Fields

operand1

operand2

save_variable

Mod

Modulo operation: save_variable = operand1 % operand2.

Fields

operand1

operand2

save_variable

Or

Bitwise OR: save_variable = operand1 | operand2.

Fields

operand1

operand2

save_variable

And

Bitwise AND: save_variable = operand1 & operand2.

Fields

operand1

operand2

save_variable

Not

Bitwise NOT: result = ~operand.

Fields

operand

result

Jump

Jump to a label.

Fields

jump_to_label

LoadW

Loads a word: variable = array_address[index].

Fields

array_address

index

variable

SetCursor

Positions the cursor at the specified line and column.

Fields

line

col

SetCursorOperand

Positions the cursor at the line and column in the given Operands.

Fields

row

col

UpdateCursorPos

Update the cursor position variable with current data.

GetCursor

Store the current cursor position at store_addr.

Fields

store_addr

EraseWindow

Erase the entire window with the specified id.

Fields

value

EraseLine

Erase the current line starting from the cursor.

SetVarType

Changes the variable type of the specified variable.

Fields

variable

vartype

CopyVarType

Copies the variable type of from to variable.

Fields

variable

from

GetVarType

Stores the variable type of variable in result.

Fields

variable

result

Quit

Quits the Z-Machine interpreter immediately.