Struct zwreec::backend::zcode::zfile::Zfile

pub struct Zfile {
    pub data: Bytes,
    pub object_addr: u16,
    pub type_store: u16,
    pub cursor_pos: u16,
    pub heap_start: u16,
    pub bright_mode: bool,
    pub force_unicode: bool,
    pub easter_egg: bool,
    pub no_colours: bool,
    pub no_unicode: bool,
    // some fields omitted
}

The definition of a Z-Code file.

Fields

data

The output data

object_addr

Location of object table

type_store

Location of the type storage

cursor_pos

Location of the cursor position

heap_start

Start of dynamic memory

bright_mode

Flag to enable black font on white background

force_unicode

Force print_unicode op-code generation and omit unicode-translation table generation

easter_egg

Enable the easter-egg

no_colours

Disable colours

no_unicode

Disable unicode completely

Methods

impl Zfile

fn new() -> Zfile

Creates a new zfile with default options.

fn new_with_options(bright_mode: bool, force_unicode: bool, easter_egg: bool, no_colours: bool, half_memory: bool, no_unicode: bool) -> Zfile

Creates a new zfile with the specified options.

fn new_with_cfg(cfg: &Config) -> Zfile

Creates a new zfile with the specified config.

fn create_header(&mut self)

Creates the header of a zfile.

fn write_string(&mut self, newstring: &str) -> u16

Saves the string to high memory.

fn add_jump(&mut self, name: String, jump_type: JumpType)

Adds a jump to write the jump-addresses after reading all commands.

fn emit(&mut self, code: Vec<ZOP>)

Write out the ZOP instructions to the data.

fn write_zop(&mut self, instr: &ZOP, return_new_jumps: bool) -> (Vec<Zlabel>, Vec<Zjump>, Vec<u8>)

Write opcodes to data array but also return written bytes for testing purposes as well as the resulting new labels and jumps.

fn gen_print_ops(&mut self, text: &str)

Generates normal print opcodes for ASCII characters and unicode print opcodes for unicode characters. Adds new characters to the unicode translation table if there is still space.

fn start(&mut self)

Start of a zcode program.

Fills everything < program_addr with zeros.

Caution

This should be called as the first command.

fn end(&mut self)

Writes all stuff that couldn't be written directly.

Caution

This should be called as the last command.

fn routine(&mut self, name: &str, count_variables: u8)

Command to create a Z-Routine.

fn label(&mut self, name: &str)

Command to create a label.

fn routine_add_link(&mut self)

Routine to add the address of a passage-link.

fn op_quit(&mut self)

Exits the program immediately.

quit is 0OP

fn routine_check_links(&mut self)

Checks all stored links and make them selectable with the keyboard.

The routine checks if there are < 10 links or more:

if < 10: number keys 1-9 are supported, jumps immediately.

if >=10: 99 links are supported, leading zeroes are not allowed. To jump to a link with a number smaller than 10 you have to press enter.

fn routine_check_more(&mut self)

Easter-egg, with konami-code to start.

fn routine_print_unicode(&mut self)

Print UTF-16 string at addr.

Expects an address as argument where the first u16 stored is the length of the string as the number of u16 chars, followed by the string to print. This only works if the string is within the address space up to 0xffff.

fn update_cursor_pos(&mut self)

Update the cursor position in the global cursor_pos variable.

fn routine_prompt(&mut self)

Needed to simulate a javascript browser input dialog, receives a prompt message and a default value as string arguments.

fn routine_malloc(&mut self)

malloc Z-Routine: Allocate a specified number of words of dynamic memory.

argument: amount of u16 to allocate

After receiving the address you are requested to write down the number of u16 you are actually using in the first u16 and then if you ever want to decrease this, you have to write -1i16 at the 'freed' u16s at the end. increasing it is not allowed. memory will be freed after each passage if there is no global variable pointing to it.

fn routine_strcpy(&mut self)

strcpy Z-Routine: Copy a string.

first argument is pointer to utf16 string containing length at first u16 second the the destination address in memory where the string is copied to, while the first length u16 is not copied.

fn routine_strcat(&mut self)

strcat Z-Routine: Concatenate two strings.

returns a reference to a string concatenation of the first and second string parameters.

fn routine_strcmp(&mut self)

strcmp Z-Routine: Compare two strings.

returns 0 if both given strings are equal and -1 if the first is alphabetically smaller than the second and +1 vice versa.

fn routine_malloc_init(&mut self)

malloc_init Z-Routine: Initialize the dynamic memory.

fn routine_mem_free(&mut self)

mem_free Z-Routine: Free unused dynamic memory.

This is implemented as a simple tracing garbage collector.

fn routine_manual_free(&mut self)

manual_free Z-Routine: manual free call to erase used heap memory if you can not wait for the GC.

fn routine_itoa(&mut self)

itoa Z-Routine: Convert an int to a string.

convert from number at arg1 to string at base of 10, returns the str addr.

fn routine_print_var(&mut self)

helper function to print out the content of a variable according to its type.

fn routine_add_types(&mut self)

Helper function to add two values according to the types of them and saves type of savevarid to the global type-store and returns the result.

fn routine_print_char(&mut self)

Print one zscii character given as argument.

fn op_jump(&mut self, jump_to_label: &str)

Jumps to a label.

fn op_call_1n(&mut self, jump_to_label: &str)

Calls a routine.

call_1n is 1OP.

fn op_call_2n_with_address(&mut self, jump_to_label: &str, address: &str)

Calls a routine with an argument(variable) and throws result away because the value isn't known until all routines are set, it inserts a pseudo routoune_address.

call_2n is 2OP.

fn op_call_2n_with_arg(&mut self, jump_to_label: &str, arg: &Operand)

Calls a routine with one argument an throws result away.

call_2n is 2OP.

fn op_call_2s(&mut self, jump_to_label: &str, arg: &Operand, result: &Variable)

Calls a routine with one argument and stores return value in result.

call_2s is 2OP.

fn op_call_vn_a2(&mut self, jump_to_label: &str, arg1: &Operand, arg2: &Operand)

Calls a routine with two arguments and throws result away.

call_vn is VAROP.

fn op_call_vn_a3(&mut self, jump_to_label: &str, arg1: &Operand, arg2: &Operand, arg3: &Operand)

Calls a routine with three arguments and throws result away.

call_vn is VAROP.

fn op_call_vs_a2(&mut self, jump_to_label: &str, arg1: &Operand, arg2: &Operand, result: &Variable)

Calls a routine with two arguments and stores return value in result.

call_vs is VAROP.

fn op_call_vs_a3(&mut self, jump_to_label: &str, arg1: &Operand, arg2: &Operand, arg3: &Operand, result: &Variable)

Calls a routine with three arguments and stores return value in result.

call_vs is VAROP.

fn op_call_vs2_a5(&mut self, jump_to_label: &str, arg1: &Operand, arg2: &Operand, arg3: &Operand, arg4: &Operand, arg5: &Operand, result: &Variable)

Calls a routine with five arguments and stores the return value.

call_vs2 is VAROP with additional types-byte.

fn op_je(&mut self, operand1: &Operand, operand2: &Operand, jump_to_label: &str)

Jumps to a label if the value of operand1 is equal to operand2.

fn op_jne(&mut self, operand1: &Operand, operand2: &Operand, jump_to_label: &str)

Jumps to a label if the value of operand1 is not equal to operand2.

fn op_jl(&mut self, operand1: &Operand, operand2: &Operand, jump_to_label: &str)

Jumps to a label if the value of operand1 is lower than operand2 (compared as i16).

fn op_jle(&mut self, operand1: &Operand, operand2: &Operand, jump_to_label: &str)

Jumps to a label if the value of operand1 is lower than or equal operand2 (compared as i16).

fn op_jge(&mut self, operand1: &Operand, operand2: &Operand, jump_to_label: &str)

Jumps to a label if the value of operand1 is greater than or equal operand2 (compared as i16).

fn op_jg(&mut self, operand1: &Operand, operand2: &Operand, jump_to_label: &str)

Jumps to a label if the value of operand1 is greater than operand2.

fn op_read_char_timer(&mut self, local_var_id: u8, timer: u8, routine: &str)

Reads keys from the keyboard and saves the asci-value in local_var_id.

read_char is VAROP.

fn op_print_unicode_char(&mut self, character: u16)

Prints a unicode char to the current stream.

fn op_print_unicode_var(&mut self, variable: &Variable)

Prints variable with unicode char to the current stream.

fn op_print_unicode_str(&mut self, address: &Operand)

Prints a unicode string to the current output stream.

fn op_print_char(&mut self, variable: &Variable)

Prints a ZSCII character.