Struct zwreec::backend::zcode::zfile::Zstring

pub struct Zstring {
    pub from_addr: u32,
    pub chars: Vec<u8>,
    pub orig: String,
    pub unicode: bool,
    pub written_addr: u32,
}

A string.

Fields

from_addr

the address where the string address should be stored at

chars

The character data Contains either ztext or [length: u16, utf16char:u16, …]

orig

The original string

unicode

Contains whether the string is a unicode string or ZSCII

written_addr

If the string data was already written to a location this is the address