Module zwreec::config

Configuration for the zwreec compiler.

Uses the getopts library to append a getopts::Options with compiler specific flags and then creates a Config struct from getopts::Matches.

Usage

This module depends on the getopts crate. To use it, you need to add getopts to your project. Read getopts' Usage for more information.

Example

The following example shows how to generate a Config struct from an empty getopts::Options.

extern crate zwreec;
extern crate getopts;

use zwreec::config;
use zwreec::config::Config;

fn main() {
    let args: Vec<String> = std::env::args().collect();
    let opts = config::zwreec_options(getopts::Options::new());

    let matches = match opts.parse(&args[1..]) {
        Ok(m) => m,
        Err(f) => { panic!(f.to_string()) }
    };

    let cfg = Config::from_matches(&matches);
}

Appending this module (Development)

To add new compiler flags for the compiler, you need to change three parts in this module.

Appending the Struct

To store your new setting, you need to add it as a field to the Config struct. Here we add a new boolean pub italics: bool to control italics. Always add a field-comment.

pub struct Config {
    /// Add easter egg to compiler
    pub easter_egg: bool,
    /// Instruct compiler to run these test-cases
    pub test_cases: Vec<String>,
    /// Enables or disables Italics
    pub italics: bool,
}

Now you also need to change Config::default_config() to provide the default value for your new field:

pub fn default_config() -> Config {
    Config{
        easter_egg: true,
        test_cases: Vec::new(),
        italics: true,
    }
}

Parse your flag in from_matches()

from_matches uses the Options -F and -N to enable or disable boolean flags. Therefore, if you added a new boolean flag, you can now simply add it to the match statement.

for s in matches.opt_strs("F") {
    match s.as_ref() {
        "easter-egg" => {
             cfg.easter_egg = true;
             debug!("enabled easter-egg");
        },
        "italics" => {
            cfg.italics = true;
            debug!("enabbled italics");
        },
        _ => {
            error!("Cannot enable feature {} - feature not known.", s);
        }
    }
}

for s in matches.opt_strs("N") {
    match s.as_ref() {
        "easter-egg" => {
            cfg.easter_egg = false;
            debug!("disabled easter-egg");
        },
        "italics" => {
            cfg.italics = false;
            debug:("disalbed italics");
        }
        _ => {
            error!("Cannot disable feature {} - feature not known.", s);
        }
    }
}

Adding something that is not a boolean

If you need to add an option that is not a boolean, you will need to append the struct and default_config() accordingly (e.g. pub notaflag: String,). Then you need to add a new getopts::Option to zwreec_options():

pub fn zwreec_options(mut opts: getopts::Options) -> getopts::Options {
opts.optmulti("F", "feature", "", "FEAT");
opts.optmulti("N", "no-feature", "enable or disable a feature (can occur multiple times).
                    List of supported features (default):
                        easter-egg (enabled)", "FEAT");
opts.optflag("e", "generate-sample-zcode", "writes out a sample zcode file, input file is not used and can be omitted");

opts.optopt("n", "notaflag", "notaflag", "SOMETHING");

opts
}

Now you can append Config::from_matches() to analyse the provided matches for your new option and set the Config accordingly:

if let s = matches.opt_str("n") {
    cfg.notaflag = s;
}

Now you can use the new flag inside the compiler.

Structs

Config
Represents the configuration for the compiler.

Enums

TestCase
The Type used to define backend tests for the compiler.

Functions

zwreec_options
Appends a getopts::Options with compiler specific flags.

zwreec_usage
Prints a usage