Function zwreec::config::zwreec_options

pub fn zwreec_options(opts: Options) -> Options

Appends a getopts::Options with compiler specific flags.

The method Config::from_matches() looks for very specific getopts::Matches. This function takes a getopts::Options to append it with Options required by from_matches. It currently adds three fields:

opts.optmulti("F", "feature", "", "FEAT");
opts.optmulti("N", "no-feature", "enable or disable a feature (can occur multiple times).
                    List of supported features (default):
                        easter-egg (enabled)", "FEAT");
opts.optflag("e", "generate-sample-zcode", "writes out a sample zcode file, input file is not used and can be omitted");

Example

You can use this function to append your getopts::Options.

let mut opts = getopts::Options::new();
opts.optflag("h", "help", "print this message");

let opts = zwreec::config::zwreec_options(opts);

Another useful example is to use it to gernerate a more compact usage by having a function that only returns your options.

fn options() -> getopts::Options {
    let mut opts = getopts::Options::new();
    opts.optflag("h", "help", "display this help and exit");
    opts.optflag("V", "version", "display version");

    opts
}

fn print_usage(program: &str, verbose: bool) {
    let brief = format!("Usage: {} [options]", program);

    let opts = if verbose {
        zwreec::config::zwreec_options(options())
    } else {
        options()
    };

    print!("{}", opts.usage(&brief));
}

As you can see, options() returns your own command line options, which are then conditionally expanded by using zwreec_options().