Struct zwreec::config::Config

pub struct Config {
    pub bright_mode: bool,
    pub easter_egg: bool,
    pub force: bool,
    pub force_unicode: bool,
    pub half_memory: bool,
    pub no_colours: bool,
    pub no_unicode: bool,
    pub unsupported_formatting: bool,
    pub test_cases: Vec<TestCase>,
}

Represents the configuration for the compiler.

This struct is created using either config::default_config() or config::parse_matches(matches: &getopts::Matches). It contains a series of fields that can be used to control the behaviour of the compiler.

Example Usage inside the compiler

let cfg = config::Config::default_config();

if cfg.easter_egg {
    println!("Egg!");
}

Fields

bright_mode

Force a bright background and dark text

easter_egg

Add easter egg to compiler

force

Force compilation despite errors

force_unicode

Force generation of print_uncode opcodes

half_memory

Divide memory usage by 2

no_colours

Disable colours

no_unicode

Disable unicode support

unsupported_formatting

Enable Formatting Simulation

test_cases

Instruct compiler to run these test-cases

Methods

impl Config

fn default_config() -> Config

Returns a Config struct with preset defaults

Example

use zwreec::config::Config;
let cfg = Config::default_config();

fn from_matches(matches: &Matches) -> Config

Returns a Config struct by using getopts::Matches to set the fields.

This method analyses a getopts::Matches for fields provided by zwreec_options().

Example

let args: Vec<String> = std::env::args().collect();
let opts = zwreec::config::zwreec_options(getopts::Options::new());

let matches = match opts.parse(&args[1..]) {
    Ok(m) => m,
    Err(f) => { panic!(f.to_string()) }
};

let cfg = zwreec::config::Config::from_matches(&matches);