Function zwreec::compile

pub fn compile<R: Read, W: Write>(cfg: Config, input: &mut R, output: &mut W)

Compiles a Twee Input to Zcode

This is the main call into the Zwreec compiler. It will process input: Read by calling the different parts of the compiler-chain, notably frontend::lexer to generate a tokenstream, frontend::parser and frontend::ast to generate the Abstract Syntax Tree and lastly frontend::codegen to generate the Zcode.

Example

let cfg = zwreec::config::Config::default_config();
let mut input = File::open(Path::new(&args[1])).unwrap();
let mut output = File::create(Path::new("a.z8")).unwrap();

zwreec::compile(cfg, &mut input, &mut output);