Function zwreec::test_library

pub fn test_library<R: Read, W: Write>(cfg: Config, input: &mut Option<R>, output: &mut Option<W>)

Run internal library tests.

This function is used to circumvent certain parts of the compiler toolchain. It currently only processes TestCase::ZcodeBackend which creates a Zcode file using all available OP-Codes.

Warning: This function should be considered unstable and might be removed in later versions.

Example

let mut cfg = zwreec::config::Config::default_config();
cfg.test_cases.push(zwreec::config::TestCase::ZcodeBackend);

let mut input: Option<File> = None;
let mut output = Some(File::create(Path::new("a.z8")).unwrap());

zwreec::test_library(cfg, &mut input, &mut output);