Module zwreec::frontend

Processes Twee files and builds an abstract syntax tree.

The frontend is responsible for processing a Twee file by generating a Token Stream from the input and parsing this stream to build an Abstract Syntax Tree.

The first stage in generating the AST is the lexer. The tokens generated by the lexer are then fed to the parser. The parser generates a sequence of operations that are then used to generate the AST. The AST can then be parsed to generate code in the backend.

Example

This small example shows how the different submodules interact and can be used to parse a Twee-File.

use std::io::Cursor;

let cfg = zwreec::config::Config::default_config();
let mut twee = Cursor::new("::Start\nHello World".to_string().into_bytes());

// Clean Input
let mut cursor = zwreec::frontend::screener::handle_bom_encoding(&mut twee);

// Generate Token Stream
let tokens = zwreec::frontend::lexer::lex(cfg.clone(), &mut cursor);

// Parse Tokens
let p = zwreec::frontend::parser::Parser::new(cfg.clone());
let ast: Vec<zwreec::frontend::ast::ASTNode> = zwreec::frontend::ast::ASTBuilder::build(cfg, p.parse(tokens)).collect();

Modules

ast
The ast module contains a lot of useful functionality to create and walk through the ast (abstract syntaxtree)

evaluate_expression
The evaluate_expressions module evaluates expressions from the AST and compiles them into Z-Code.

expressionparser
The expressionparser module parses every expression to an AST (abstract syntax tree). The idea is explained here: http://programmers.stackexchange.com/questions/254074/

lexer
Builds a token iterator for twee input.

parser
Constructs a parsing iterator for a twee token iterator.

screener
Sanitizes the input stream.