Module zwreec::frontend::ast

The ast module contains a lot of useful functionality to create and walk through the ast (abstract syntaxtree)

Structs

ASTBuilder
This is the state of the AST building operation.

NodeDefault
This is a default node.

NodePassage
This is a passage node.

Enums

ASTNode
This is a generic node.

ASTOperation
The parser uses these ASTOperations to create the AST.