Enum zwreec::frontend::ast::ASTNode

pub enum ASTNode {
    Default(NodeDefault),
    Passage(NodePassage),
}

This is a generic node.

It can either be a standard node or a passage node.

Variants

Default

A standard node

Passage

Passage node

Methods

impl ASTNode

The implementation of a node.

fn add_child(&mut self, path: Vec<usize>, token: Token)

Adds a child to the path in the AST

fn count_childs(&self, path: Vec<usize>) -> usize

Counts the childs of the current path in the AST.

fn is_specific_token(&self, token: Token, path: Vec<usize>) -> bool

Checks the current path if there is the token specified by token.

fn category(&self) -> Token

Returns the current category (Token) of a node.

fn childs(&self) -> &Vec<ASTNode>

Returns all child trees of a node.

fn as_default(self) -> NodeDefault

Unwraps the ASTNode as NodeDefault.

Panics

Panics if the ASTNode is not a NodeDefault.