Enum zwreec::frontend::ast::ASTOperation

pub enum ASTOperation {
    AddPassage(Token),
    AddChild(Token),
    ChildDown(Token),
    ChildUp(Token),
    TwoUp,
    Up,
    UpChild(Token),
    UpChildDown(Token),
    UpSpecial,
}

The parser uses these ASTOperations to create the AST.

Variants

AddPassage

Adds a passage to the path in the ast.

AddChild

Adds a child to the path in the ast.

ChildDown

Adds a child and adds the child to the current path.

ChildUp

Adds a child and goes one level up.

TwoUp

Goes two levels up.

Up

Goes one level up in the path.

UpChild

Goes one level up and adds a child.

UpChildDown

Goes one level up, adds a child and adds the added child to the path.

UpSpecial

This goes one level up or goes out of an if-expression if possible.