Struct zwreec::frontend::ast::AST

pub struct AST {
    // some fields omitted
}

Methods

impl AST

fn build<I: Iterator<Item=ASTOperation>>(ops: I) -> AST

fn operation(&mut self, op: ASTOperation)

fn add_passage(&mut self, token: Token)

adds a passage to the path in the ast

fn add_child(&mut self, token: Token)

adds a child to the path in the ast

fn child_down(&mut self, token: Token)

adds a child an goees one child down

fn two_childs_down(&mut self, child1: Token, child2: Token)

adds one child and goes down. adds snd child and goes down.

fn up(&mut self)

goes one lvl up

fn up_special(&mut self)

special up of the if-expression

fn up_child(&mut self, token: Token)

goes one lvl up and adds and child

fn up_child_down(&mut self, token: Token)

goes one lvl up, adds an child and goes one lvl down

fn two_up(&mut self)

goes two lvl up

fn to_zcode(&self, out: &mut Zfile)

convert ast to zcode

fn print(&self, force_print: bool)

prints the tree

fn count_childs(&self, path: Vec<usize>) -> usize

counts the childs of the path in the asts

fn is_specific_token(&self, token: Token, path: Vec<usize>) -> bool

checks in the ast if there is the token "token"