Struct zwreec::frontend::ast::ASTBuilder

pub struct ASTBuilder {
    // some fields omitted
}

This is the state of the AST building operation.

Methods

impl ASTBuilder

fn build<I: Iterator<Item=ASTOperation>>(cfg: Config, ops: I) -> Scan<Constructor<Peeking<I, ASTOperation>, ASTBuilder, ASTNode, fn(&mut ASTBuilder, &mut Option<ASTNode>, (ASTOperation, Option<ASTOperation>)) -> Option<ASTNode>>, Config, fn(&mut Config, ASTNode) -> Option<ASTNode>>

Constructs an iterator over all passages.

fn operation(&mut self, current_passage: &mut Option<ASTNode>, op: ASTOperation) -> Option<ASTNode>

Calls the matching function to a given ASTOperation.

fn count_childs(&self, current_passage: &mut Option<ASTNode>, path: Vec<usize>) -> usize

Counts the childs of the path in the AST.

fn add_passage(&mut self, token: Token) -> Option<ASTNode>

Adds a passage to the path in the AST.

fn add_child(&mut self, current_passage_opt: &mut Option<ASTNode>, token: Token) -> Option<ASTNode>

Adds a child to the path in the AST.

fn child_down(&mut self, current_passage: &mut Option<ASTNode>, token: Token) -> Option<ASTNode>

Adds a child and adds the child to the current path.

fn two_childs_down(&mut self, current_passage: &mut Option<ASTNode>, child1: Token, child2: Token) -> Option<ASTNode>

Adds one child and goes down. Adds a second child and goes down.

fn up(&mut self) -> Option<ASTNode>

Goes one level up in the path.

fn up_special(&mut self) -> Option<ASTNode>

This goes one level up or goes out of an if-expression if possible.

fn child_up(&mut self, current_passage: &mut Option<ASTNode>, token: Token) -> Option<ASTNode>

Adds a child and goes one level up.

fn up_child(&mut self, current_passage: &mut Option<ASTNode>, token: Token) -> Option<ASTNode>

Goes one level up and adds a child.

fn up_child_down(&mut self, current_passage: &mut Option<ASTNode>, token: Token) -> Option<ASTNode>

Goes one level up, adds a child and adds the added child to the path.

fn two_up(&mut self) -> Option<ASTNode>

Goes two levels up.