Module zwreec::frontend::evaluate_expression

The evaluate_expressions module evaluates expressions from the AST and compiles them into Z-Code.

This module takes the root node of an expression and traverses the contained subtree. It analyses if a sub-expression only contains constants and if so, evaluates them while compiling. Otherwise it translates the expressions into Z-Code. The module uses a finite amount of local variables in Z-Code to evaluate the expressions. Hence only expressions with limited size are supported.

Enums

EvaluateExpressionError
All the possible errors that can occur during parsing.

Functions

evaluate_expression
This functions evaluates an expression from the AST and returns an Operand containing the result.