Enum zwreec::frontend::evaluate_expression::EvaluateExpressionError

pub enum EvaluateExpressionError {
    InvalidAST,
    NumericOperatorNeedsTwoArguments {
        op_name: String,
        location: (u64, u64),
    },
    UnhandledToken {
        token: Token,
    },
    UnsupportedOperator {
        op_name: String,
        location: (u64, u64),
    },
    UnsupportedFunction {
        name: String,
        location: (u64, u64),
    },
    UnsupportedFunctionArgsLen {
        name: String,
        location: (u64, u64),
        expected: u64,
    },
    UnsupportedFunctionArgType {
        name: String,
        index: u64,
        location: (u64, u64),
    },
    NoTempIdLeftOnStack,
}

All the possible errors that can occur during parsing.

Variants

InvalidAST

Some unexpected Node in the AST

NumericOperatorNeedsTwoArguments

A numerical operator always needs two arguments

Fields

op_name

location

UnhandledToken

Unsupported Token

Fields

token

UnsupportedOperator

Unsupported operator

Fields

op_name

location

UnsupportedFunction

Unsupported function

Fields

name

location

UnsupportedFunctionArgsLen

The count of the function args is wrong / unexpected

Fields

name

location

expected

UnsupportedFunctionArgType

The type of the function arg is wrong / unexpected

Fields

name

index

location

NoTempIdLeftOnStack

Expression is too complex