Module zwreec::frontend::lexer

Builds a token iterator for twee input.

This is lexical analyser using rustlex to analyse twee input and create a lazy evaluating iterator over lexical tokens.

About Rustlex

From rustlex' readme:

RustLex is a lexical analysers generator, i.e. a program that generate lexical analysers for use in compiler from a description of the language using regular expressions.

The Zwreec development team forked rustlex to enable support for some advanced features like preserving accurate line and column numbers for tokens and fix some limitations of the original project. You can find the fork on Github at Farthen/rustlex.

For more information about Rustlex, visit the original repository by LeoTestard.

The TweeLexer Struct

The TweeLexer Struct and its accompaning implementations are autogenerated by rustlex' compiler plugin and cannot be documented using rustdoc. To get a better impression of how it is supposed to work, take a look at the uncompiled source code

Structs

ScanState
Stores the state for the custom iterator scan_filter().

TweeLexer

Enums

LexerError
The errors that can occur in the lexer.

Token
The resulting Tokens that are returned by the lex function.

Functions

lex
Builds a Token iterator for twee input.