Enum zwreec::frontend::lexer::Token

pub enum Token {
    TokPassage {
        location: (u64, u64),
        name: String,
    },
    TokTagStart {
        location: (u64, u64),
    },
    TokTagEnd {
        location: (u64, u64),
    },
    TokVarSetStart {
        location: (u64, u64),
    },
    TokVarSetEnd {
        location: (u64, u64),
    },
    TokPassageLink {
        location: (u64, u64),
        display_name: String,
        passage_name: String,
    },
    TokTag {
        location: (u64, u64),
        tag_name: String,
    },
    TokText {
        location: (u64, u64),
        text: String,
    },
    TokFormatBoldStart {
        location: (u64, u64),
    },
    TokFormatBoldEnd {
        location: (u64, u64),
    },
    TokFormatItalicStart {
        location: (u64, u64),
    },
    TokFormatItalicEnd {
        location: (u64, u64),
    },
    TokFormatUnderStart {
        location: (u64, u64),
    },
    TokFormatUnderEnd {
        location: (u64, u64),
    },
    TokFormatStrikeStart {
        location: (u64, u64),
    },
    TokFormatStrikeEnd {
        location: (u64, u64),
    },
    TokFormatSubStart {
        location: (u64, u64),
    },
    TokFormatSubEnd {
        location: (u64, u64),
    },
    TokFormatSupStart {
        location: (u64, u64),
    },
    TokFormatSupEnd {
        location: (u64, u64),
    },
    TokFormatMonoStart {
        location: (u64, u64),
    },
    TokFormatMonoEnd {
        location: (u64, u64),
    },
    TokFormatBulList {
        location: (u64, u64),
    },
    TokFormatNumbList {
        location: (u64, u64),
    },
    TokFormatIndentBlock {
        location: (u64, u64),
    },
    TokFormatHorizontalLine {
        location: (u64, u64),
    },
    TokFormatHeading {
        location: (u64, u64),
        rank: u8,
        text: String,
    },
    TokMacroStart {
        location: (u64, u64),
    },
    TokMacroEnd {
        location: (u64, u64),
    },
    TokMacroContentVar {
        location: (u64, u64),
        var_name: String,
    },
    TokMacroSet {
        location: (u64, u64),
    },
    TokMacroIf {
        location: (u64, u64),
    },
    TokMacroElse {
        location: (u64, u64),
    },
    TokMacroElseIf {
        location: (u64, u64),
    },
    TokMacroEndIf {
        location: (u64, u64),
    },
    TokMacroPrint {
        location: (u64, u64),
    },
    TokMacroDisplay {
        location: (u64, u64),
        passage_name: String,
    },
    TokMacroSilently {
        location: (u64, u64),
    },
    TokMacroEndSilently {
        location: (u64, u64),
    },
    TokMacroNoBr {
        location: (u64, u64),
    },
    TokMacroEndNoBr {
        location: (u64, u64),
    },
    TokParenOpen {
        location: (u64, u64),
    },
    TokParenClose {
        location: (u64, u64),
    },
    TokVariable {
        location: (u64, u64),
        name: String,
    },
    TokArrayAccess {
        location: (u64, u64),
        name: String,
        index: String,
    },
    TokArrayLength {
        location: (u64, u64),
        name: String,
    },
    TokInt {
        location: (u64, u64),
        value: i32,
    },
    TokFloat {
        location: (u64, u64),
        value: f32,
    },
    TokString {
        location: (u64, u64),
        value: String,
    },
    TokBoolean {
        location: (u64, u64),
        value: String,
    },
    TokFunction {
        location: (u64, u64),
        name: String,
    },
    TokColon {
        location: (u64, u64),
    },
    TokArgsEnd {
        location: (u64, u64),
    },
    TokArrayStart {
        location: (u64, u64),
    },
    TokArrayEnd {
        location: (u64, u64),
    },
    TokAssign {
        location: (u64, u64),
        var_name: String,
        op_name: String,
    },
    TokNumOp {
        location: (u64, u64),
        op_name: String,
    },
    TokCompOp {
        location: (u64, u64),
        op_name: String,
    },
    TokLogOp {
        location: (u64, u64),
        op_name: String,
    },
    TokSemiColon {
        location: (u64, u64),
    },
    TokNewLine {
        location: (u64, u64),
    },
    TokUnaryMinus {
        location: (u64, u64),
    },
    TokExpression,
    TokError {
        location: (u64, u64),
        message: String,
    },
}

The resulting Tokens that are returned by the lex function.

These Tokens are matched by our lexical analyser. Every Token but TokExpression stores their original location inside the analysed input. Some Tokens also use other fields to store additional information.

Variants

TokPassage
Fields

location

name

TokTagStart
Fields

location

TokTagEnd
Fields

location

TokVarSetStart
Fields

location

TokVarSetEnd
Fields

location

TokPassageLink
Fields

location

display_name

passage_name

TokTag
Fields

location

tag_name

TokText
Fields

location

text

TokFormatBoldStart
Fields

location

TokFormatBoldEnd
Fields

location

TokFormatItalicStart
Fields

location

TokFormatItalicEnd
Fields

location

TokFormatUnderStart
Fields

location

TokFormatUnderEnd
Fields

location

TokFormatStrikeStart
Fields

location

TokFormatStrikeEnd
Fields

location

TokFormatSubStart
Fields

location

TokFormatSubEnd
Fields

location

TokFormatSupStart
Fields

location

TokFormatSupEnd
Fields

location

TokFormatMonoStart
Fields

location

TokFormatMonoEnd
Fields

location

TokFormatBulList
Fields

location

TokFormatNumbList
Fields

location

TokFormatIndentBlock
Fields

location

TokFormatHorizontalLine
Fields

location

TokFormatHeading
Fields

location

rank

text

TokMacroStart
Fields

location

TokMacroEnd
Fields

location

TokMacroContentVar
Fields

location

var_name

TokMacroSet
Fields

location

TokMacroIf
Fields

location

TokMacroElse
Fields

location

TokMacroElseIf
Fields

location

TokMacroEndIf
Fields

location

TokMacroPrint
Fields

location

TokMacroDisplay
Fields

location

passage_name

TokMacroSilently
Fields

location

TokMacroEndSilently
Fields

location

TokMacroNoBr
Fields

location

TokMacroEndNoBr
Fields

location

TokParenOpen
Fields

location

TokParenClose
Fields

location

TokVariable
Fields

location

name

TokArrayAccess
Fields

location

name

index

TokArrayLength
Fields

location

name

TokInt
Fields

location

value

TokFloat
Fields

location

value

TokString
Fields

location

value

TokBoolean
Fields

location

value

TokFunction
Fields

location

name

TokColon
Fields

location

TokArgsEnd
Fields

location

TokArrayStart
Fields

location

TokArrayEnd
Fields

location

TokAssign
Fields

location

var_name

op_name

TokNumOp
Fields

location

op_name

TokCompOp
Fields

location

op_name

TokLogOp
Fields

location

op_name

TokSemiColon
Fields

location

TokNewLine
Fields

location

TokUnaryMinus
Fields

location

TokExpression

TokError
Fields

location

message

Methods

impl Token

fn location(&self) -> (u64, u64)

Returns the original location inside the twee input as (line, column).

impl Token

fn is_same_token(&self, other: &Token) -> bool

Tests is two tokens are of the same type.

We re-create tokens with location {0,0} in the parser Thus comparing also works by ignoring the content of the token.