Function zwreec::frontend::lexer::lex

pub fn lex<R: Read>(cfg: Config, input: R) -> FilteringScan<Peeking<TweeLexer<BufReader<R>>, Token>, ScanState, fn(&mut ScanState, (Token, Option<Token>)) -> Option<Token>>

Builds a Token iterator for twee input.

This function takes a twee input and uses the TweeLexer generated by rustlex to create a token iterator. The tokens are then cleaned using the Peeking iterator to provide lookahead and the FilteringScan iterator to merge adjacent text tokens and combine variable assignment.

Example

use std::io::Cursor;

let cfg = zwreec::config::Config::default_config();
let mut input = Cursor::new("::Start\nHello World".to_string().into_bytes());

let tokens = zwreec::frontend::lexer::lex(cfg, &mut input);