Module zwreec::frontend::parser

Constructs a parsing iterator for a twee token iterator.

This is a predictive parser for a twee token stream. It takes a Token iterator (see lexer) and wraps it inside a parsing iterator that returns operations to construct an abstract syntax tree.

Parser

The underlying parser is a predictive top-down parser for a LL(1) grammar. For more information about the parser you should refer to Compilers: Principles, Techniques, and Tools by A. V. Aho, M. S. Lam, R. Sethi, and J. D. Ullman, Chapter 4.4.4: "Nonrecursive Predictive Parsing".

Grammar

As mentioned, the parser operates on a LL(1) grammar. It is documented in Zwreec's Github Wiki, together with the resulting parse table

Structs

ParseState
Stores the stack for the custom iterator parsing().

Parser
The Parser parses the Token stream and generates AST operations.

Enums

Elem
The Type that represents an element of the grammar.

NonTerminalType
The Type of nonterminal encountered by the parser.

ParserError
All the possible errors that can occur during parsing.