Enum zwreec::frontend::parser::Elem

pub enum Elem {
    NonTerminal(NonTerminalType),
    Terminal(Token),
}

The Type that represents an element of the grammar.

This enum represents the lexical elements of a grammar. Nonterminals are defined by NonTerminalType, terminals use the lexical token as specification.

For the full grammar, take a look at Zwreec's Wiki.

Variants

NonTerminal

A non-terminal element

Terminal

A terminal element