Enum zwreec::frontend::parser::ParserError

pub enum ParserError {
    TokenDoNotMatch {
        token: Option<Token>,
        stack: Token,
    },
    StackIsEmpty {
        token: Token,
    },
    NoProjection {
        token: Token,
        stack: NonTerminalType,
    },
    NonTerminalEnd {
        stack: NonTerminalType,
    },
}

All the possible errors that can occur during parsing.

Variants

TokenDoNotMatch

Token does not match the token on the stack

Fields

token

stack

StackIsEmpty

The stack is empty but there are still Tokens left

Fields

token

NoProjection

No projection exists for the current Token to the non-terminal on the stack

Fields

token

stack

NonTerminalEnd

Ending at a non-terminal

Fields

stack