Module zwreec::utils::error

Implementation of error logging for errors thrown in the compiler

This module contains Display implementations for the errors defined in the core modules (frontend/backend), allowing to log them easily. Also contains macros to log and panic - depending on the Config module - because of an error, making the actual error handling code in the other modules much shorter and easier to use.