Module zwreec::utils::extensions

Extensions to existing types used in the zwreec compiler

Currently this contains only extensions to the Iterator trait yielding new Iterator types that wrap certain functionality, that is used to filter/construct/lookahead all objects passed and iterated upon in the compiler, making it possible to have most of the compiler chain to be lazy_evaluated and even multi-threaded.

Structs

Cached
An Iterator that performs all iteration in a seperate therad and caches them

Constructor
An iterator to maintain state while iterating another iterator and allows to hold the currently active object to modify it (/construct it), until the next element is created

FilteringScan
An iterator to maintain state while iterating another iterator and allows filtering

Parser
An iterator to maintain state while iterating another iterator and allows more complex parsing

Peeking
An iterator that performs a lookahead of 1, utilizing the existing Peakable Iterator Can be stacked to perform an even greater lookahead

Enums

ParseResult
The result of a parsing operation

Traits

ConstructorExt
An extension to sized Iterators that specifies a function that returns a Constructor Iterator

FilteringScanExt
An extension to sized Iterators that specifies a function that returns a FilteringScan Iterator

ParserExt
An extension to sized Iterators that specifies a function that returns a Parser Iterator

PeekingExt
An extension for sized Iterators that specifies a function that returns a Peeking Iterator

Functions

cached
A function to create a cached Iterator by passing the old Iterator by closure. If the underlying iterator does not support Send you can construct it in the closure