Struct zwreec::backend::zcode::zfile::Zfile

pub struct Zfile {
    pub data: Bytes,
    pub object_addr: u16,
    pub heap_start: u16,
    pub force_unicode: bool,
    pub easter_egg: bool,
    // some fields omitted
}

Fields

data

object_addr

heap_start

force_unicode

easter_egg

Methods

impl Zfile

fn new() -> Zfile

creates a new zfile

fn new_with_options(force_unicode: bool, easter_egg: bool) -> Zfile

fn new_with_cfg(cfg: &Config) -> Zfile

fn create_header(&mut self)

creates the header of a zfile

fn write_string(&mut self, newstring: &str) -> u16

fn add_jump(&mut self, name: String, jump_type: JumpType)

adds jump to write the jump-addresses after reading all commands

fn emit(&mut self, code: Vec<ZOP>)

write out respective byte stream of opcodes to file

fn write_zop(&mut self, instr: &ZOP) -> (Vec<Zlabel>, Vec<Zjump>, Vec<u8>)

write opcodes to file but also return written bytes for testing purposes as well as the resulting new labels and jumps

fn gen_print_ops(&mut self, text: &str)

generates normal print opcodes for ASCII characters and unicode print opcodes for unicode characters

fn start(&mut self)

start of an zcode programm fills everying < program_addr with zeros should called as the first commend

fn end(&mut self)

writes all stuff that couldn't written directly should be called as the last commend

fn routine(&mut self, name: &str, count_variables: u8)

command to create a routine

fn label(&mut self, name: &str)

command to create a label

fn routine_add_link(&mut self)

routine to add the address of a passage-link

fn op_quit(&mut self)

exits the program quit is 0OP

fn routine_check_links(&mut self)

checks all stored links and make them choiceable with the keyboard

fn routine_check_more(&mut self)

easter-egg, with konami-code to start

fn routine_print_unicode(&mut self)

print UTF-16 string from addr expects an address as argument where first the length (as u16) of the string is written (number of u16 chars), followed by the string to print. only works if the string is within the address space up to 0xffff

fn routine_malloc(&mut self)

malloc

fn routine_strcpy(&mut self)

strcpy

fn routine_strcat(&mut self)

strcat returns a reference to a string concatenation of the first and second string parameters

fn routine_malloc_init(&mut self)

malloc_init

fn routine_mem_free(&mut self)

mem_free as a tracing garbage collection

fn routine_itoa(&mut self)

itoa convert from number at arg1 to string at base of 10, returns the str addr

fn op_jump(&mut self, jump_to_label: &str)

jumps to a label

fn op_call_1n(&mut self, jump_to_label: &str)

calls a routine call_1n is 1OP

fn op_call_2n_with_address(&mut self, jump_to_label: &str, address: &str)

calls a routine with an argument(variable) an throws result away becouse the value isn't known until all routines are set, it inserts a pseudo routoune_address call_2n is 2OP

fn op_call_2n_with_arg(&mut self, jump_to_label: &str, arg: &Operand)

calls a routine with one argument an throws result away call_2n is 2OP

fn op_call_2s(&mut self, jump_to_label: &str, arg: &Operand, result: &Variable)

calls a routine with one argument and stores return value in result call_2s is 2OP

fn op_call_vn_a2(&mut self, jump_to_label: &str, arg1: &Operand, arg2: &Operand)

calls a routine with two arguments and throws result away call_vn is VAROP

fn op_call_vn_a3(&mut self, jump_to_label: &str, arg1: &Operand, arg2: &Operand, arg3: &Operand)

calls a routine with three arguments and throws result away call_vn is VAROP

fn op_call_vs_a2(&mut self, jump_to_label: &str, arg1: &Operand, arg2: &Operand, result: &Variable)

calls a routine with two arguments and stores return value in result call_vs is VAROP

fn op_call_vs_a3(&mut self, jump_to_label: &str, arg1: &Operand, arg2: &Operand, arg3: &Operand, result: &Variable)

calls a routine with two arguments and stores return value in result call_vs is VAROP

fn op_je(&mut self, operand1: &Operand, operand2: &Operand, jump_to_label: &str)

jumps to a label if the value of operand1 is equal to operand2

fn op_jne(&mut self, operand1: &Operand, operand2: &Operand, jump_to_label: &str)

jumps to a label if the value of operand1 is not equal to operand2

fn op_jl(&mut self, operand1: &Operand, operand2: &Operand, jump_to_label: &str)

jumps to a label if the value of operand1 is lower than operand2 (compared as i16)

fn op_jle(&mut self, operand1: &Operand, operand2: &Operand, jump_to_label: &str)

jumps to a label if the value of operand1 is lower than or equal operand2 (compared as i16)

fn op_jge(&mut self, operand1: &Operand, operand2: &Operand, jump_to_label: &str)

jumps to a label if the value of operand1 is greater than or equal operand2 (compared as i16)

fn op_jg(&mut self, operand1: &Operand, operand2: &Operand, jump_to_label: &str)

jumps to a label if the value of operand1 is greater than operand2

fn op_read_char_timer(&mut self, local_var_id: u8, timer: u8, routine: &str)

reads keys from the keyboard and saves the asci-value in local_var_id read_char is VAROP

fn op_print_unicode_char(&mut self, character: u16)

prints an unicode char to the current stream

fn op_print_unicode_var(&mut self, variable: &Variable)

prints variable with unicode char to the current stream

fn op_print_unicode_str(&mut self, address: &Operand)