Module zwreec::utils::extensions

Structs

FilteringScan
An iterator to maintain state while iterating another iterator and allows filtering

Parser
An iterator to maintain state while iterating another iterator and allows more complex parsing

Peeking
An iterator that performs a lookahead of 1, utilizing the existing Peakable Iterator Can be stacked to perform an even greater lookahead

Enums

ParseResult

Traits

FilteringScanExt

ParserExt

PeekingExt